moving all dialog related files to a new created subdirectory "dialogs"
