+++ qtodotxt/ui/dialogs/about_view.py
