+++ qtodotxt/ui/dialogs/task_editor_service.py
@@ -1,5 +1,5 @@
 import string
-from qtodotxt.ui.controls.autocomplete_inputdialog import AutoCompleteInputDialog
+from qtodotxt.ui.dialogs.autocomplete_inputdialog import AutoCompleteInputDialog
 from collections import OrderedDict
